Remove requested from SelectorState
A SelectorState represents the state required to resolve a concrete ComponentIdentifier instance from
a ComponentSelector. Previously, SelectorState instances would be identified by the requested
component selector -- the selector that was originally requested as part of a dependency.
During selector resolution, we would use that requested component selector as the one we
attempt to resolve.

This does not accurately reflect the intention of a selector state, which is to resolve
the actual component selector after dependency substitutions are performed. The only reason
we track the original component selector before substitutions is to provide that data
in a ResolutionResult, where we expose the original requested selector and the final
resolved component. A requested selector is a property of an EdgeState/DependencyState. The
SelectorState should only be concerned with resolving the final substituted component selector.

This commit solifies this concept by removing the requested component selector from the
SelectorState entirely. In addition it removes the DependencyState field from SelectorState
as well, as that dependency state represents the requested selector. Now, selectors are keyed
on the substituted component selector instead of the requested.

When visiting an assembled graph, we used the visitSelector method to visit all requested component
selectors, as a SelectorState previously was incorrectly based on a the requested selector. Now that
they are properly based on the substituted selector, visiting SelectorState instances will no longer
allow us to visit requested selectors in the final result.

To resolve this, we no longer visit all selectors when building a ResolutionResult. Instead, we get
the requested component selector from the edge. However, this means we lose the previous deduplication
and caching that we achieved given that SelectorStates were cached based on the requested selector.
Instead, we deduplicate requested selectors at the time of serialization. We introduce a new
DeduplicatingComponentSelectorSerializer that does this.

SelectorState instance are still likely tied up into too much logic not related to resolving a
component selector into a concrete component identifier, however this commit is an
initial first step into cleaning up the concept of a selector