Cache Artifact Transforms across projects
In many cases, projects are initialized with functionally identical artifact transform registrations.
While the registrations themselves may have different parameters, they are initialized in the same order
and have the same from/to attributes.

Similar to caching we do with variant selection, and caching already done for source artifact sets in the transform selector,
we can store the indices of input artifact transforms, effectively reusing transform selection results across projects.

This commit creates a cacheable representation of a set of transform registrations for a given project, and then uses that
as an input to the artifact transform selection process. This cacheable representation is mapped back to the concrete
transform registrations after results are loaded from the cache.

This results in a notable reduction of artifact selection time in projects that use artifact transforms heavily.