Rename UnlockedConfiguration to LegacyConfiguration
Use the legacy terminology publicly to mean any configuration with mutable roles.
This conflits with our existing definition of legacy, which meant any configuration with all three usages enabled.
Some internal renaming needed to be done to avoid overloading the legacy terminology.

There is still some existing usages of Unlocked in the ConfigurationContainer.
The rework was avoided here, as we will be touching all the Unlocked methods in 9.0 when migrating away from legacy configurations.