Verify that the number of entries written matches the written number
It is theoretically possible for a collection to change the number of
entries while being traversed. It might or might not result in a
ConcurrentModificationException, but the resulting issues can then only
be discovered when reading and are really hard to trace back.

There is no need to check when reading, because read always relies on
the written size to determine the number of elements.